Check for parameter existence (see #6600)
Description
-----------

We need to check for this parameter to exist. Unfortunately, Symfony itself does not provide a `hasParameter()` but `getParameter()` only which throws a `ServiceNotFoundException` in case there was no `parameter_bag` service registered. So one would always need to `try/catch` this. I figured, a `hasParameter()` can be useful for us.

Commits
-------

c229b4c0 Check for parameter existence