Remove column from articles URL (see #6594)
Description
-----------

While preparing for the _Content URL Generator_, I noticed this: did you know that (sometimes) Contao generates the URL to an article as `.../articles/section:alias` and not just `.../articles/alias`?

The reason behind this is, that `Controller::getFrontendModule` will ignore the articles parameter if it does not belong the the given column (section). **However**, this was

a) not applied in all place
b) totally unnecessary because we can achieve the same correct thing without having it in the URL (see PR code)
c) allowed for URL manipulation, I could theoretically generate an article in the wrong section (not saying that would be a real problem, but … )

This PR removes the need for passing the section in the article alias, because it was irrelevant to most places anyway. This makes generating the URL of an article easier and probably also fixes bugs because no third-party developer I know would currently have generated a correct URL 😅

Commits
-------

6713bfee Remove column from articles URL
d4f75595 Also adjust the `Controller::redirectToFrontendPage()` method

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>