Fix toggle operation parameter check (see #6659)
Description
-----------

I noticed a very strange issue while working on article voters. Currently, if a user does not have access to `tl_article::published`, the toggle button is not shown at all. After implementing a voter instead of the current callback, the eye button was visible if disallowed, but not visible if allowed (= the article can be edited, but the field iself is not allowed).

I traced this down to the `DataContainer` special handling for the toggle button. The `DataContainer` automatically hides the toggle button completely in a DCA, if the user does not have access to the toggling field. However, if a button is disabled, the DC does not know what field the toggle operation is about (because there's no `href`), and does nothing 😂 

This PR fixes this by falling back to the original href defined in the DCA operation, even if it is removed in the button callback.

Commits
-------

376e5b18 Fix toggle operation parameter check