Fetch visible root trail record from database (see #6670)
Description
-----------

Rendering visible root trails in a tree is a very special case. Because the user actually does not have read permissions on the given record.

We currently use `DataContainer::getCurrentRecord` to fetch visible trail pages. However, if correct permissions are implemented (e.g. on tl_page), the user would not have access to that page. It would therefore not render this node, and therefore not render any nodes of any trees (because the root records are not readable).

I first tried to add a `$checkPermissions` property to `getCurrentRecord`, but that does not really work with our cache of either records or exception.

Commits
-------

da34fcd3 Fetch visible root trail record from database
dc4bd1a0 Improve logic and added comment
ac273c42 Correctly fetch the current record