Normalize the Twig CSP method names (see #6759)
Description
-----------

As discussed on Slack this normalises the Twig CSP method names.

The reason why the nonce method was called `contao_csp_nonce` and not `csp_nonce` originally is that `nelmio/security-bundle` provides its own `csp_nonce` method. This PR renames the `contao_csp_nonce` method now to `csp_nonce`, overwriting the `csp_nonce` method from Nelmio - but forwarding to Nelmio's `CSPRuntime` if the method is called without a Contao response context. This also requires the minimum version of `nelmio/security-bundle` to be raised to `^3.0`, since the `CSPRuntime` service will not exist otherwise.

Commits
-------

3ce20d73 normalise CSP twig method names
1409209e fix composer.json