Fix routes with parameters in the SERP widget (see #6839)
Description
-----------

Currently, the back end would show an exception when trying to edit a page with parameters. The widget previously handled routing errors and did show a `no preview available` message. The new exception handling when generating the URL was added for if a route is generated for a content that does not have a content route provider. But we must not handle the route parameter exception here.

Commits
-------

edbb15b3 Fix route with parameters in the Serp Widget