Add `type="button"` to the accordion toggler (see #6898)
Description
-----------

If you happen to insert an accordion content element in a form, e.g. in an HTML code form field with the `{{insert_content::*}}` insert tag, the toggler of the accordion will currently submit the form when clicked, instead of just opening/closing the accordion.

This is because any `<button>` without a specific `type` will be treated as a `type="submit"` button by default: https://html.spec.whatwg.org/multipage/form-elements.html#the-button-element

> The attribute's missing value default and invalid value default are both the _Submit Button state_.

Since our toggler button is not supposed to submit anything but rather just act "as a button", it needs `type="button"`.

Commits
-------

c5e8687f add type attribute to accordion toggler