Fix the Twig loader infrastructure (see #6936)
Description
-----------

This PR fixes a problem where templates would not be recognized when clearing the cache in production.  
Spoiler alert: :rabbit::hole: 

### The original problem
In the existing implementation, we had a mandatory cache warmer, the `ContaoFilesystemLoaderWarmer`, that gathered the Contao template paths, added them to our loader and made it persist its cache. Symfony's `cache:clear` is a command, so when running it on an empty/nuked cache, it first builds the container and thus executing the mandatory cache warmers. After that the command finally clears the cache and runs the cache warmers gain. However, now, only the optional ones, reasoning the others had already run when the container was built. So our custom cache, living in the cache pool, gets killed (and not rebuild). :man_facepalming:  It was only a side-effect, that it still worked in dev mode (auto-refresh logic).

So yeah, this is a design issue due to mixed up responsibilities. It resulted from the attempt to build a loader similar to Twig's own. But it turned out, that ours is quite different, because it **a)** must know about individual template files and **b)** must allow a refresh at runtime.

### The fix
* The responsibility to find and memorize templates is now moved to the `ContaoFilesystemLoader` in its entirety. That means, there are no more functions to build, clear or persist the inheritance chains. They will get created (and cached) on demand or, optionally, when calling a new `warmUp()` method. This makes so much more sense as it encapsulates the state and inner workings and just provides the hierarchy every time someone requests it.

* The `ContaoFilesystemLoaderWarmer` is now an *optional cache warmer* and basically only calls `ContaoFilesystemLoader#warmUp()`. I also extracted the listener that automatically refreshed the template hierarchy in dev mode to its own class because this was also a mix of responsibilities.

* As the `ContaoFilesystemLoader` is the only thing ever, that knows about the template hierarchy, I also dropped the `TemplateHierarchyInterface`. This was an attempt to separate concerns but keeping it would now add even more confusion than before. If you need the template hierarchy, inject the `ContaoFilesystemLoader`. Btw.: there are quite some files changed in this PR, however, most of the changes outside the loader/cache warmer files are just switching `TemplateHierarchyInterface` with `ContaoFilesystemLoader`.

* Moving things into the `ContaoFilesystemLoader` class also made it obvious, that extending from Twig's filesystem loader is a ~~bad~~ unbelievably stupid idea (also see #6871). I therefore removed the parent class, inlined the calls and refactored a lot of code. Without the `parent::` calls, there is now also less filesystem access, because - surprise - our loader already knows about all the files when creating the hierarchy.

### 4.13
This PR is targeted against 5.3 because it was much simpler to do (and still a lot of pain). If things look good, we should probably backport it to 4.13 as well (maybe dropping the tests).  


/cc @bytehead, @fritzmg, @contaoacademy

*Please help testing this PR in your projects to ensure all the template hierarchy related features still work. This includes rendering legacy templates with Twig, extending and horizontally reusing things and so on...*

Commits
-------

74af88b4 rework the ContaoFilesystemLoader infrastructure
a87997ff adjust tests