Disable the search index listener in the back end (see #7148)
Description
-----------

Accessing the `$document->getContentCrawler()`  is a very heavy operation if you have `masterminds/html5` installed because Symfony's DomCrawler component will use this parser automatically to parse the entire HTML.
This can be very heavy, especially in the back end. I noticed that this listener is not needed in the back end at all, it's only wasted resources.

I didn't bother fixing it in 4.13 because I don't use it anywhere anymore and by default, Contao does not have that problem because it does not require `masterminds/html5` itself. You will only have this useless slow-down when having it installed.

Commits
-------

a2f0e646 Skip search index listener in back end