Correctly resolve child definitions for callback tags (see #7169)
Description
-----------

I was trying to add another `imgSize` field to `tl_module` and wanted to re-use the `contao.listener.image_size_options` listener which would be super easy like so:

```php

    $services->set('my_listener')
        ->parent('contao.listener.image_size_options')
        ->tag('contao.callback', [
            'table' => 'tl_module',
            'target' => 'fields.my_image_sizes_field.options',
        ])
    ;
```

However, this currently fails because parent services are not correctly resolved. This PR fixes this.
The `while (!$definition->getClass() && $definition instanceof ChildDefinition) {` logic is Symfony standard and can be found in many other places for example the `RegisterEventListenersAndSubscribersPass` which has bascially more or less exactly the same responsibility as our `DataContainerCallbackPass` 😊 

I've also updated the `TestListener` which technically is not necessary. I could've used another method to mock this logic but I figured I wanted to work with my use case which as the options listener 😉

Commits
-------

97da131b Fixed missing child definition support for callback tags
d00c19e4 Both services must be public