Check access to `fieldsOfTable` for the file edit operation (see #7175)
Description
-----------

Fixes https://github.com/contao/contao/issues/7038

The permission check is applied by the `TableAccessVoter` here. It is intentionally only applied to the edit operation, because move or copy should not check for `canEditFieldsOfTable`.

Also does not work with Contao 4.13 – not sure we want to backport that. We don't check any such permissions in 4.13. Contrary to 5.3, where an access denied exception is thrown by the `TableAccessVoter`, in Contao 4.13 we just render an empty list of fields to edit everywhere.

---

Side note: I also started working on actual file permissions in https://github.com/contao/contao/compare/5.3...aschempp:contao:fix/files-permissions. There are a number of unknowns though, e.g. is editing a file (content) not the same as editing file metadata. It also related to the virtual file system and a possible redesign of the file manager by @m-vo. After discussing stuff like the `UploadAction` I postponed that for now and setteled for this "quick fix".

Commits
-------

d9843ba3 Check fieldsOfTable access for file edit operation