Skip permissions checks for child records (see #7133)
Description
-----------

This restores the permission handling we had in all previous Contao versions. After discussing with @Toflar, this is the only viable solution to fix our permission issues. Here are two examples that are imho impossible to fix otherwise. They are not a technical issue, but a logical one.

1. If you duplicate a page, its articles will be duplicated as well as its content elements. If you do not have access to some elements, the whole page and articles are duplicated, **but "random" content elements will be missing**.

2. even worse, when deleting a page, DC_Table tries to delete all child records and create an undo record for them. The ones that the user does not have access will not be deleted, **and will not be added to the undo record**. They are however still removed by our reviseTable, and will therefore be lost without the option to undo.

In the same sense we could discuss to adjust the `undo` method instead of what we discussed in https://github.com/contao/contao/pull/7056

FYI @Toflar and me discussed that the same handling will not be possible in a regular REST/CRUD API. It would require special APIs for that, which can always be added if necessary.


- Fixes https://github.com/contao/contao/issues/7101
- Probably also makes https://github.com/contao/contao/pull/7102 obsolete but correctly checking the IDs/updating the cache might still be correct.

Commits
-------

eb388c66 Skip permissions checks for child records
961033f3 Also remove permission check in undo operation
70d75771 Fixed duplicate variable name