Only handle `GET` requests in the search index listener (see #7240)
Description
-----------

I noticed that `$this->indexer->delete($document);` is still executed for `POST` requests - if the response has a failure status code. Thus it is currently possible to remove pages from the index via `POST` requests, if you can make the `POST` request fail - which is easy, as you just have to send a cookie with the request and omit the request token.

I noticed this via spam attacks that invoked error responses (since there were so many request, a deadlock error occured during `$this->indexer->delete($document)`).

The _"Should be skipped because it is not a GET request"_ test did not pick up on this as it only tested for a sucessful response. I have adjusted the test accordingly. Or should we split it up into two tests instead? One with a succesful response and one with an unsuccesful response.

Commits
-------

595ceaad skip non-GET requests outright
268c67b8 adjust comment