Overwrite the page metadata before parsing the news article (see #7228)
Description
-----------

This PR fixes the fact that the page metadata is overwritten after the content has been generated. The sequence is relevant, for example, if an insert tag is used in the content that outputs the page metadata, e.g. `{{page::pageTitle}}`.
If the page metadata is overwritten afterwards the insert tag will be replaced with the page title set in tl_page, with this PR the insert tag is replaced with the title of the news, as it should be IMO.

Commits
-------

00d2dd77 Overwrite page metadata before parsing the news article