Make ID always searchable in DC_Table (see #7275)
Description
-----------

Problem:

1. Go to https://demo.contao.org/contao?do=form
2. Try to search for `Form ID` 3
3. You won't get any results.

The reason for that is that `Form ID` relates to the HTML ID and not the database ID. This is confusing because e.g. in `tl_article` or `tl_page` you can search for the database ID so you expect to be able to do that in forms as well.

I think the easiest fix for that problem is to allow searching for the database ID for form related stuff as well.

Commits
-------

00d1ae11 Fix searching for form data IDs
7cafcc0a Make ID always searchable in DC_Table