Store the back end popup session bag under a different storage key (see #7273)
Description
-----------

This is an alternative approach to #7221 and #7269. It registers an additional session bag for the popup, under a different storage key, but under the same `contao_backend` name. This way you can still use the `contao_backend` session bag everywhere and don't have to worry about whether your back end module was opened in the popup or not. And you don't need an additional service to find the correct bag.

Caveat: this will not store the session state of the popup in the database (`tl_user.session`) anymore (same as in #7269 as far as I can see). So once you log out and then back in, all your filters etc. that were used in the popup only will be gone. But I think that's reasonable. However, if we don't want this, I can still change the `UserSessionListener` so that is stores both states (regular and popup).

Commits
-------

6a826575 automatically store contao_backend session attributes for the popup u…
2c62a775 only check for popup once in UserSessionListener
5557c1bf add tests
01b8ae84 code style
d6dbcb5a hard code session bags
03ceab0c Clean up and add the missing unit test (#9)

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>