Fix the login state for two factor authentication (see #7323)
Description
-----------

Fixes #6265, #7272

As discussed in the Contao call on 2024-07-04 the solution is fairly simple. By treating logged in users that haven't fully authenticated themselves via 2FA yet (i.e. when the token is a `TwoFactorTokenInterface`, as also checked by the `AuthenticationSuccessHandler`) as guests, both issues are solved.

#6265 is fixed, because the 401 page will be rendered again after the login - and the login module on the 401 page will then show the 2FA form.

#7272 is fixed because only the pages (and content) visible to guests will be shown.

Commits
-------

7040f0f9 fix 2FA (#6265, #7272)
abd03a0b use interface