Handle routing errors when trying to render the error screen (see #7332)
Description
-----------

If a domain is mapped to a Contao installation, but it has no root page for it, that currently throws an internal server error instead of a 404 page. Because our pretty error screen renderer tries to render a 404 page for the current domain, but can't find its root page.

Commits
-------

a9279007 Handle routing errors when trying to render the error screen