Ignore deleted cookies in the `CsrfTokenCookieSubscriber` (see #7350)
Description
-----------

Fixes #7344

This PR ignores any cookies that are about to be deleted in the response while checking the request and response cookies.

This PR is based against 4.13. While the original issue is directly observable in Contao 5.3 with the `AutoExpiringAttribute`, the same issue can also occur in 4.13.

Commits
-------

0f4b4bdb ignore deleted cookies
d708cdcc cs fix
3a822522 add test
b2180ba9 remove superfluous use statement