Add a Stimulus controller to configure TinyMCE instances (see #7394)
Description
-----------

Fixes one part of #7391

This PR introduces a Stimulus controller for the tinyMCE editor. This way we can gracefully remove and (re-)init editors independent of the document structure.

I tried to keep the changes to the `be_tinyMCE` template to a minimum. If you're only overwriting blocks, nothing needs to be adjusted. The only difference is that we're building a config that's passed on to the controller instead of directly instantiating the editor.

An alternative to passing a JSON string would be to store the config in a global variable as there can only be one at a time anyways (currently).

Commits
-------

d5aeaab1 add a stimulus controller to configure tiny mce instances
99b99b1b use automatic JSON parsing and proper disconnect
def09222 Rebuild the assets

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>