Revert 'Move the `assets` folder to `public/assets`' (see #7406)
Description
-----------

Unfortunately, I have to declare the `public/assets` attempt a failure 😞 

The asset URLs are now incorrectly generated as `/public/assets/…` instead of `/assets/…` which we didn‘t notice because the images route was adjusted accordingly. But that is wrong, of course.

#7400 shows all the problems that occur when trying to fix this.

At the moment, neither the image studio nor `contao/image` are prepared to handle URL prefixes. We can pass a `$rootDir` parameter to methods like `getImg()`, `getSources()` and `getImageSrc()`, but we will always end up with wrong paths:

* `href="/../files/lightbox-images/…"` and `src="/assets/images/…"` or
* `href="/files/lightbox-images/…"` and `src="/public/assets/images/…"`

So this PR reverts the changes for now.

If we really want to go down this path, we first have to adjust the image studio and `contao/image` to support different file paths and URL paths.

Commits
-------

0d391ff8 Revert 'Move the `assets` folder to `public/assets`'