Introduce a general cache tag invalidator service (see #7566)
Description
-----------

Our cache invalidation logic currently operates directly on the HTTP cache only which sometimes is not sufficient because when a user e.g. invalidates a content element, you may also want to invalidate some API cache information or whatever.

This has to be done on Contao level and not in a general HttpCache(Bundle) event because it's logically incorrect there. The HttpCache(Bundle) operates on the HTTP cache. What we want here is a general tag invalidation. Invalidating the HTTP cache tags is **one** of the steps but thanks to this new event, it can be many more.

Commits
-------

f8ec5e60 Introduce a general CacheTagInvalidator service
f69dbbab Update core-bundle/src/Event/InvalidateCacheTagsEvent.php