Add the buttons builder (see #7645)
Description
-----------

This refactors the button building into a new (maybe temporary) class. It shows a lot of duplicate code in the DCs. I intentionally left the markup the same to make it reviewable, we can then refactor the `ButtonsBuilder` to a template in a next step.

There are few differences now "thanks" to unified code:
 1. We correctly check for sorting field in extended tree and parent mode. Only if sorting is enabled we now offer the "copy repeatedly" and "cut" operation. This was not the case previously, which I think was kindof a bug.
 2. In DC_Folder, the "cut" operation was before "copy". This has changed due to unified buttons in the PR. I don't see why that should be a problem :) 
 3. "copyRepeatedly" was not available in DC_Folder, it has now been added.
 4. The `nb` ("no back") URL attribute was not taken into account in a few places to hide the "save and close" button

Commits
-------

62af6c76 Added buttons builder
02715e67 Fix copyMultiple in DC_Folder
d4fe0460 Extract all DC_Folder buttons