Use the VFS in the `FileProvider` of the back end search (see #7627)
Description
-----------

This is a first draft how the VFS could be used in the backend search's `FileProvider` instead of directly accessing `tl_files`.

IMHO we do not need to gather any metadata when indexing because it can just be retrieved when creating the hit (no need to serialize stuff that is not meant to be serialized and so on). 

I also changed the way images are added - now we're adding pre-configured `FigureBuilder` instances to the hit (a bit similar to the strategy in the file previews). Later on either the controller or the template can then build the figures and render them.

/cc @Toflar

Commits
-------

5324fd6b use the VFS in the FilesProvider
d202a205 rename FilesProvider to FilesStorageProvider
9411f591 add another test for the permission checking vfs