Move the bundle templates into the `@Contao` namespace (see #7659)
Description
-----------

Fixes #7568 

This PR moves the bundle templates into the `@Contao` namespace. In order not to break things, I opted to leave the existing `@ContaoCore/Error/*` and `@ContaoCore/Frontend/*` templates in place and added an `extends` statement to the new ones. When we're ready to drop these in Contao 6, we can then adjust the Controllers to directly render from the managed namespace. This is also why I added the deprecation notice as a comment instead of using the `{% deprecated %}` tag (we still need to call them).

There is a small fix in the preview toolbar: The icon visibility was completely broken (probably introduced with #7586), which is now fixed by setting the fill to `#fff` again in that scenario.

It might be a good idea to review file names of the new template files (`@Contao`). I tried to clean things up a bit and make them consistent. :slightly_smiling_face:

Commits
-------

db6b7d02 move and temporarily rewire the @ContaoCore/Error/* templates
e6d880de move the @ContaoCore/Backend/* templates
32cd47d6 move the @ContaoCore/Collector/* templates
b0fb8333 move and temporarily rewire the @ContaoCore/Frontend/* templates (fro…
127ddbd7 also add a deprecation notice to the @ContaoCore/Image/* templates
c2f00da0 disallow adjusting templates from the "frontend_preview" and "web_deb…
196dc1d5 fix comment
64f54a55 rename backend_fallback to backend