Add the domain to the "root page dependent module" configuration (see #7732)
Description
-----------

I currently have multiple domains in an installation, where the root page is always named after the language. This results in this useless view:

<img width="586" alt="Bildschirmfoto 2024-11-20 um 17 20 22" src="https://github.com/user-attachments/assets/4fbce17e-a45e-4276-b5b7-c5bff7f77ce9">

This PR adds the domain name to the dropdown option. It should be cherry-pickable to Contao 5.3, or do we need a separate PR?

Commits
-------

dc72da6e Add the domain to the RootPageDependentModule config