Register folders, pages and nodes as tooltip targets (see #7841)
Description
-----------

### Description

* Fixes #7840 

This registers pages, folders and the toggle node buttons as a tooltips-target.

### Reason

The afterload method only triggers once after the controller has been initialized. However with this, we'll have new targets that are connected so they'll work on Ajax.
They won't be initialized twice as the tooltips-controller checks for the data-attribute when migrating the legacy titles.

Commits
-------

43f43738 Register folders, pages and nodes as tooltip target