Pass on the information if a message has been triggered via the WebWorker (see #7814)
Description
-----------

This passes on the information whether the current message is handled within the `WebWorker` or not. 

I have played with multiple different approaches to see whether it should be an information on the `Envelope` instead of the message and work with a custom middleware that would skip those messages. Then I also tried using an attribute on the message but I realized that all of those approaches prevent the message handler from having the power to do things dynamically. 
For example, you might not just want to skip a message but instead delay it for 5 minutes waiting if there will be real workers then. Or you want to check something yourself in the message handler to determine whether it actually can be done within a web request (e.g. count some data and detect if it's lower than `n` it should be possible or whatever).

So it is part of the message now, giving full control to the message handler - also it's super easily testable for message handler tests etc.

Commits
-------

e01607c5 Pass on info if a message was triggered using the WebWorker
19f9fdde Use scope as a string to have the DX identical to cron jobs
c8dcba7b Fixed tests