Allow setting a custom redirect route when unauthenticated (see #7834)
Description
-----------

Currently, when the back end session expires (or the cookie is removed otherwise), a redirect to the login mask happens. We are adding a request parameter `redirect` to allow redirecting back to the last requested URL. This works fine until the requested URL is not supposed to be visited in the browser, which is the case for every Turbo stream request or any generic Ajax request. 

**Example:** Your session has expired and you try to open the text content element in the template studio. Your redirect URL will now be `/contao/template-studio/resource/content_element/text`, which is a route requiring a specific content type and thus will produce a 404.

This PR introduces a new `_unauthenticated_redirect_route` request attribute. When set, the `ContaoLoginAuthenticator` uses this route for the `redirect=…` request parameter. This allows controllers to specify a dedicated fallback route, which I made use of in the `BackendTemplateStudioController`.

Commits
-------

54ebae34 introduce a _unauthenticated_redirect_route request parameter, that a…
22180a36 use the _unauthenticated_redirect_route parameter in the template stu…