Fix the `tooltips-controller` (see #7873)
Description
-----------

Currently the tooltips do not always work when a page is rendered by Turbo Drive. The issue is that the Mutation Observer only checks the list of added nodes, but not any of their children.

Basically the same fix as [here](https://github.com/contao/contao/pull/7824#discussion_r1905774862).

Commits
-------

624cc51e Fix tooltips-controller