Do not use `DC_Table::generateRecordLabel()` for breadcrumbs (see #7939)
Description
-----------

Fixes #7870

This should be a more robus way to generate the title. It does no longer instantiate a `DC_Table` object so the issues from #7870 should be solved with that.

This also fixes #7887

It uses the `$GLOBALS['TL_DCA'][$table]['list']['sorting']['defaultSearchField']` as the default tilte and if this is not set or empty we fallback to the `tl_$table.edit` or `DCA.edit` label.

Commits
-------

ca506915 Do not use DC_Table::generateRecordLabel() for breadcrumbs