Remove background from regular `MODE_PARENT` listings (see #7942)
Description
-----------

If you have a very simple DCA that uses `MODE_PARENT`

```php
$GLOBALS['TL_DCA']['tl_foobar'] = [
    'config' => [
        'dataContainer' => DC_Table::class,
        'ptable' => …,
        'sql' => …,
    ],
    'list' => [
        'sorting' => [
            'mode' => DataContainer::MODE_PARENT,
            'fields' => …,
            'headerFields' => …,
        ],
        'label' => [
            'fields' => …,
            'format' => '%s',
        ],
        'operations' => [],
    ],
    'fields' => …,
];
```

The list will be currently rendered with a grey background:

<img src="https://github.com/user-attachments/assets/bc2cc6d8-d14b-4ff5-a569-e85622847a94" width="383">

Contrary to previous Contao versions, where such a list would have no such background. Only the border between the records.

The culprit is this CSS:

https://github.com/contao/contao/blob/e5eb31bbab2f534ba542a7f0dd502e52df9b85c1/core-bundle/contao/themes/flexible/styles/main.css#L1442-L1444

As discussed on Slack, it is unclear for which use case this style was added, as it is never used in the core (at least we could not find any usages). Because either the `.as-grid` class is present - or no `<ul>` is rendered in the first place.

Commits
-------

ac708e8e Remove background for regular MODE_PARENT listings