Set `defaultSearchField` for `tl_page` to `title` (see #7988)
Description
-----------

The default search field for `tl_page` is currently set to the non-mandatory `pageTitle` field. It should be set to the mandatory `title` field instead, as that is the title shown in the back end and also the title you most likely want to search for, especially if you only ever need to fill the `pageTitle` for a small subset of pages.

This then also fixes #7987.

Commits
-------

f66fabf2 Set default search field for tl_page to title