Add missing `ref` for edit profile redirect (see #7986)
Description
-----------

Fixes #7965

As mentioned by @ausi in https://github.com/contao/contao/issues/7965#issuecomment-2625722429 this issue also occurs in 5.3 (and 4.13) if you open any other link in a new tab before clicking on _Profile_ in the header menu. This is because that profile link will actually then redirect you elsewhere, using the wrong `ref` parameter.

Additionally I changed the link in the header menu so that it does not cause a redirect and instead passes all necessary parameters to the URL generator in the first place, which also fixes the issue.

Commits
-------

a3c7b127 Add current reference for profile link
8b718d7d Use correct parameters for profile link in the first place
d01ec7bb Update test