Use a prefix for the passkey user handle (see #8021)
Description
-----------

This prefixes the `userHandle` for the passkeys (WebAuthn credentials) with `backend.`. This is done in preparation of https://github.com/contao/contao/pull/8011. The database ID of `tl_member` and `tl_user` won't be unique, so there can be situations where a passkey for the front end cannot be generated if there already exists a passkey for the back end under the same `userHandle` (i.e. user ID).

This in turn also backports a few other improvements and changes from #8011, so that they will be more forward compatible.

_Note:_ this will cause any existing Passkeys you have for your `5.5` RC instance to not work anymore. You will also need to execute

```
contao:migrate --no-interaction --with-deletes
```

as the `userType` field will be removed from the credential table, as it is not needed anymore.

Commits
-------

1038f4b6 Use a backend prefix for the passkey user handle
64ebb6a8 revert comment change for now
518cea6a Update core-bundle/contao/classes/FrontendUser.php
0b762f56 Rename the method

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>